CLN: Defer to super() as much as possible in Trapping
Change base class from SR3 to ConstrainedSR3
